move code to check content based on PR comments
